Center of Thrust .1
Added black marker to show center of thrust. Currently uses the native
KSP center of thrust calculator, so it does not take into account
current thrust, only "max" thrust, as it would in the VAB or SPH. Center
of thrust is correct even if you using a non-root part for control.
Currently proof of concept.

Changes previewed with a line of slashes for easy disambiguation.